requirements.txt
absl-py==0.8.0
alabaster==0.7.12
apiclient==1.0.4
asn1crypto==0.24.0
astor==0.8.0
astroid==1.6.6
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==2.1.3
boto==2.48.0
boto3==1.7.46
botocore==1.10.46
bz2file==0.98
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
cloudpickle==1.2.2
colorama==0.3.9
configparser==4.0.2
cryptography==2.1.4
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
decorator==4.3.0
defusedxml==0.6.0
dill==0.2.8.2
Django==1.11.27
django-cors-headers==3.0.2
djangorestframework==3.9.4
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
gast==0.3.2
gensim==3.4.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
html5lib==1.0.1
httplib2==0.13.1
idna==2.7
imagesize==1.1.0
ipaddress==1.0.17
ipykernel==4.8.2
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.21
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
jsonfield==2.0.2
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
lazy-object-proxy==1.4.2
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
mock==3.0.5
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
notebook==5.5.0
numpy==1.14.5
numpydoc==0.9.1
oauth2client==3.0.0
oauthlib==2.1.0
packaging==19.2
pandas==0.23.1
pandocfilters==1.4.2
parso==0.2.1
pathlib==1.0.1
pathlib2==2.3.4
pexpect==4.6.0
pickleshare==0.7.4
Pillow==6.2.2
plac==0.9.6
preshed==1.0.0
prometheus-client==0.7.1
prompt-toolkit==1.0.15
protobuf==3.9.1
psutil==5.6.3
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycrypto==2.6.1
pyflakes==2.1.1
Pygments==2.2.0
pygobject==3.26.1
pylint==1.9.5
pyparsing==2.2.0
pyrsistent==0.15.4
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2019.3
pyxdg==0.25
PyYAML==5.1.2
pyzmq==17.0.0
QtAwesome==0.6.0
qtconsole==4.3.1
QtPy==1.9.0
regex==2017.4.5
requests==2.19.1
requests-oauthlib==1.0.0
rope==0.14.0
rsa==4.0
s3transfer==0.1.13
scandir==1.10.0
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.9.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
smart-open==1.5.7
snowballstemmer==1.9.1
spacy==2.0.11
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.2
spyder-kernels==0.5.2
subprocess32==3.5.4
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==6.10.2
toolz==0.9.0
tornado==5.0.2
tqdm==4.23.4
traitlets==4.3.2
tweepy==3.6.0
tweet-preprocessor==0.5.0
typing==3.7.4.1
ujson==1.35
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.7.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.2.1
wordcloud==1.4.1
wordnet==0.0.1b2
wrapt==1.11.2
wurlitzer==1.0.3
youtube-dl==2019.8.13
